get post_to_endpoint tests working
