+++ extract_endpoint/tests/test_post_to_endpoint.py
@@ -88,40 +88,28 @@ def check_file_in_azure(azure_service: blob.BlockBlobService,
 def test_post_stream(azure_service: blob.BlockBlobService,
                      azure_emulator_coords: azure_utils.StorageCoordinates,
                      upload_url: str,
-                     sample_stream: NamedStream,
-                     sample_stream_content: str,
-                     sample_filename: str) -> None:
+                     # sample_timestamp: str,
+                     sample_file_contents: str,
+                     sample_filenames: str,
+                     sample_zipfile: io.BytesIO) -> None:
 
-    post_return = post_to_endpoint.post_stream(stream=sample_stream, filename=sample_filename, url=upload_url)
+    post_return = post_to_endpoint.post_stream(stream=sample_zipfile, filename="filename", url=upload_url)
     assert post_return.status_code == HTTPStatus.CREATED
-    check_file_in_azure(azure_service, azure_emulator_coords, sample_filename, sample_stream_content)
+    # check_file_in_azure(azure_service, azure_emulator_coords, endpoint.TIMESTAMP_FILENAME, sample_timestamp)
+    for name, contents in zip(sample_filenames, sample_file_contents):
+        check_file_in_azure(azure_service, azure_emulator_coords, name, contents)
 
 
 @pytest.mark.usefixtures('run_endpoint')
 def test_post_stream_stdin(azure_service: blob.BlockBlobService,
                            azure_emulator_coords: azure_utils.StorageCoordinates,
                            upload_url: str,
-                           sample_stream_stdin: NamedStream,
-                           sample_stream_content: str,
-                           sample_filename: str) -> None:
-    post_return = post_to_endpoint.post_stream(stream=sample_stream_stdin, filename=sample_filename, url=upload_url)
+                           # sample_timestamp: str,
+                           sample_file_contents: str,
+                           sample_filenames: str,
+                           sample_zipfile: io.BytesIO) -> None:
+    post_return = post_to_endpoint.post_stream(stream=sample_zipfile, filename="filename", url=upload_url)
     assert post_return.status_code == HTTPStatus.CREATED
-    check_file_in_azure(azure_service, azure_emulator_coords, sample_filename, sample_stream_content)
-
-
-@pytest.mark.usefixtures('run_endpoint')
-def test_post_stream_no_filename(azure_service: blob.BlockBlobService,
-                                 azure_emulator_coords: azure_utils.StorageCoordinates,
-                                 upload_url: str,
-                                 sample_stream: NamedStream,
-                                 sample_stream_content: str,
-                                 sample_filename: str) -> None:
-    post_return = post_to_endpoint.post_stream(stream=sample_stream, filename=None, url=upload_url)
-    assert post_return.status_code == HTTPStatus.CREATED
-    check_file_in_azure(azure_service, azure_emulator_coords, sample_filename, sample_stream_content)
-
-
-@pytest.mark.usefixtures('run_endpoint')
-def test_post_stream_stdin_no_filename(upload_url: str, sample_stream_stdin: NamedStream) -> None:
-    with pytest.raises(ValueError):
-        post_to_endpoint.post_stream(stream=sample_stream_stdin, filename=None, url=upload_url)
+    # check_file_in_azure(azure_service, azure_emulator_coords, endpoint.TIMESTAMP_FILENAME, sample_timestamp)
+    for name, contents in zip(sample_filenames, sample_file_contents):
+        check_file_in_azure(azure_service, azure_emulator_coords, name, contents)
